requirements.txt
ShopifyAPI
pandas
